+++ src/utils/graphql.ts
@@ -1,0 +1,5 @@
+export function getNodesFromConnection<Node = any>(connection: {
+  readonly edges: ReadonlyArray<{ readonly node: Node }>;
+}) {
+  return connection.edges.map(edge => edge.node);
+}
